Ssh port tunnel, restore old login
