+++ package.json
@@ -11,54 +11,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.9",
-    "@angular/common": "~11.0.9",
-    "@angular/compiler": "~11.0.9",
-    "@angular/core": "~11.0.9",
-    "@angular/forms": "~11.0.9",
+    "@angular/animations": "~11.2.0",
+    "@angular/common": "~11.2.0",
+    "@angular/compiler": "~11.2.0",
+    "@angular/core": "~11.2.0",
+    "@angular/forms": "~11.2.0",
     "@angular/http": "^7.2.16",
-    "@angular/platform-browser": "~11.0.9",
-    "@angular/platform-browser-dynamic": "~11.0.9",
-    "@angular/router": "~11.0.9",
-    "@okta/okta-angular": "1.2.1",
+    "@angular/platform-browser": "~11.2.0",
+    "@angular/platform-browser-dynamic": "~11.2.0",
+    "@angular/router": "~11.2.0",
+    "@okta/okta-angular": "3.0.1",
     "body-parser": "^1.19.0",
     "chart.js": "^2.9.4",
     "component": "^1.1.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "morgan": "^1.10.0",
-    "mysql": "2.17.1",
+    "mysql": "2.18.1",
     "mysql2": "^2.2.5",
     "ngx-timeline": "5.0.0",
-    "rxjs": "~6.6.0",
+    "node-ssh-forward": "^0.7.2",
+    "rxjs": "~6.6.3",
     "sequelize": "^6.5.0",
-    "sequelize-auto": "^0.7.10",
+    "sequelize-auto": "^0.7.11",
     "sequelize-automate": "^1.2.2",
     "serve-favicon": "^2.5.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "tslib": "^2.1.0",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.7",
-    "@angular/cli": "~11.0.7",
-    "@angular/compiler-cli": "~11.0.9",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
+    "@angular-devkit/build-angular": "~0.1102.0",
+    "@angular/cli": "~11.2.0",
+    "@angular/compiler-cli": "~11.2.0",
+    "@types/jasmine": "~3.6.3",
+    "@types/node": "^14.14.27",
+    "codelyzer": "^6.0.1",
     "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.0",
+    "jasmine-spec-reporter": "~6.0.0",
+    "karma": "~6.1.1",
     "karma-chrome-launcher": "^3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-firefox-launcher": "^2.1.0",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
     "karma-junit-reporter": "^2.0.1",
     "protractor": "~7.0.0",
-    "puppeteer": "^5.5.0",
-    "ts-node": "~8.3.0",
+    "puppeteer": "^7.1.0",
+    "ts-node": "~9.1.1",
     "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
+    "typescript": "~4.1.5"
   },
   "browser": {
     "crypto": false,
