+++ src/app/newlogin/login.component.html
@@ -1,0 +1,7 @@
+
+<input type="text" name="username" id="username" [(ngModel)]="username">
+<br>
+<input type="password" name="password" id="password" [(ngModel)]="password">
+<br>
+<input type="button" value="Login" (click)="Login()">
+<br>
